fix extended rcode

+++ msg.go
@@ -690,7 +690,7 @@ func (dns *Msg) packBufferWithCompressionMap(buf []byte, compression map[string]
 		if opt == nil {
 			return nil, ErrExtendedRcode
 		}
-		opt.SetExtendedRcode(uint8(dns.Rcode))
+		opt.SetExtendedRcode(uint16(dns.Rcode))
 	}
 
 	// Convert convenient Msg into wire-like Header.
